requirements.txt
flask
moviepy
gtts
openai-whisper
googletrans==4.0.0-rc1
